Added errorThrowing mechanism that catches all errors and displays them in the pjs console
all instances of throw(...) were replaced by throwError(...)

in throwError, the error message is stored to a new var window.error
(which is inited to 0 on noError) and outputs the current pjs error
into the processing console.

This is probably not required for web only development. But really
(really) useful for mobile development where users may not have access
to a proper javascript console output.